feat: add tutorial and leaderboard
